+++ packages/node-fetch/src/Request.ts
@@ -126,7 +126,7 @@ export class PonyfillRequest<TJSON = any> extends PonyfillBody<TJSON> implements
   }
 
   clone(): PonyfillRequest<TJSON> {
-    return new PonyfillRequest(this);
+    return this;
   }
 
   [Symbol.toStringTag] = 'Request';