+++ packages/server/test/node.spec.ts
@@ -282,6 +282,27 @@ describe('Node Specific Cases', () => {
           const resJson = await response.json();
           expect(resJson.contentLength).toBe('0');
         });
+
+        it('clones the request correctly', async () => {
+          const serverAdapter = createServerAdapter(async req => {
+            const clonedReq = req.clone();
+            const textFromClonedReq = await req.text();
+            const textFromOriginalReq = await clonedReq.text();
+            return Response.json({
+              textFromClonedReq,
+              textFromOriginalReq,
+            });
+          });
+          testServer.addOnceHandler(serverAdapter);
+          const response = await fetch(testServer.url, {
+            method: 'POST',
+            body: 'TEST',
+          });
+
+          const resJson = await response.json();
+          expect(resJson.textFromClonedReq).toBe('TEST');
+          expect(resJson.textFromOriginalReq).toBe('TEST');
+        });
       });
     },
   );