+++ e2e/vercel/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @e2e/vercel
 
+## 0.0.99
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.99
+
+## 0.0.98
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.98
+
+## 0.0.97
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.97
+
 ## 0.0.96
 
 ### Patch Changes