+++ packages/server/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @whatwg-node/server
 
+## 0.9.39
+
+### Patch Changes
+
+- [`a3732a6`](https://github.com/ardatan/whatwg-node/commit/a3732a6c85dbce173b0946f70e3628c23738c799)
+  Thanks [@ardatan](https://github.com/ardatan)! - Update content-length when compressed
+
+## 0.9.38
+
+### Patch Changes
+
+- [`f7bcbea`](https://github.com/ardatan/whatwg-node/commit/f7bcbea0c0a940870efb47faeb4b64d07d990ce4)
+  Thanks [@ardatan](https://github.com/ardatan)! - Support \`Content-Encoding: none\`
+
+## 0.9.37
+
+### Patch Changes
+
+- [#1481](https://github.com/ardatan/whatwg-node/pull/1481)
+  [`481bdfd`](https://github.com/ardatan/whatwg-node/commit/481bdfd0734b6c2c70b17dccb701b068f8aa06d9)
+  Thanks [@ardatan](https://github.com/ardatan)! - New plugin to handle E2E request compression
+
+  When the client provides `Accept-Encoding` header, if the server supports the encoding, it will
+  compress the response body. This will reduce the size of the response body and improve the
+  performance of the application.
+
+  On the other hand, if the client sends `Content-Encoding` header, the server will decompress the
+  request body before processing it. This will allow the server to handle the request body in its
+  original form. If the server does not support the encoding, it will respond with
+  `415 Unsupported Media Type` status code.
+
+  `serverAdapter`'s `fetch` function handles the compression and decompression of the request and
+  response bodies.
+
+  ```ts
+  import { createServerAdapter, Response, useContentEncoding } from '@whatwg-node/server'
+
+  const serverAdapter = createServerAdapter(() => Response.json({ hello: 'world' }), {
+    plugins: [useContentEncoding()]
+  })
+  ```
+
 ## 0.9.36
 
 ### Patch Changes